Retry creation of assignment topic a few times before giving up (#3722)
* Retry creation of assignment topic a few times before giving up

* Use Action based retry mechanism

* Fix build

* Catch interrupted exception

* Fix unittest

* Added header
