+++ pulsar-functions/runtime/src/main/java/org/apache/pulsar/functions/runtime/RuntimeUtils.java
@@ -258,109 +258,4 @@ public class RuntimeUtils {
         return result.toString();
     }
 
-    public static class Actions {
-        private List<Action> actions = new LinkedList<>();
-
-        @Data
-        @Builder(toBuilder=true)
-        public static class Action {
-            private String actionName;
-            private int numRetries = 1;
-            private Supplier<ActionResult> supplier;
-            private long sleepBetweenInvocationsMs = 500;
-            private Boolean continueOn;
-            private Runnable onFail;
-            private Runnable onSuccess;
-
-            public void verifyAction() {
-                if (isBlank(actionName)) {
-                    throw new RuntimeException("Action name is empty!");
-                }
-                if (supplier == null) {
-                    throw new RuntimeException("Supplier is not specified!");
-                }
-            }
-        }
-
-        @Data
-        @Builder
-        public static class ActionResult {
-            private boolean success;
-            private String errorMsg;
-        }
-
-        private Actions() {
-
-        }
-
-
-        public Actions addAction(Action action) {
-            action.verifyAction();
-            this.actions.add(action);
-            return this;
-        }
-
-        public static Actions newBuilder() {
-            return new Actions();
-        }
-
-        public int numActions() {
-            return actions.size();
-        }
-
-        public void run() throws InterruptedException {
-            Iterator<Action> it = this.actions.iterator();
-            while(it.hasNext()) {
-                Action action  = it.next();
-
-                boolean success;
-                try {
-                    success = runAction(action);
-                } catch (Exception e) {
-                    log.error("Uncaught exception thrown when running action [ {} ]:", action.getActionName(), e);
-                    success = false;
-                }
-                if (action.getContinueOn() != null
-                        && success == action.getContinueOn()) {
-                    continue;
-                } else {
-                    // terminate
-                    break;
-                }
-            }
-        }
-
-        private boolean runAction(Action action) throws InterruptedException {
-            for (int i = 0; i< action.getNumRetries(); i++) {
-
-                ActionResult actionResult = action.getSupplier().get();
-
-                if (actionResult.isSuccess()) {
-                    log.info("Sucessfully completed action [ {} ]", action.getActionName());
-                    if (action.getOnSuccess() != null) {
-                        action.getOnSuccess().run();
-                    }
-                    return true;
-                } else {
-                    if (actionResult.getErrorMsg() != null) {
-                        log.warn("Error completing action [ {} ] :- {} - [ATTEMPT] {}/{}",
-                                action.getActionName(),
-                                actionResult.getErrorMsg(),
-                                i + 1, action.getNumRetries());
-                    } else {
-                        log.warn("Error completing action [ {} ] [ATTEMPT] {}/{}",
-                                action.getActionName(),
-                                i + 1, action.getNumRetries());
-                    }
-
-                    Thread.sleep(action.sleepBetweenInvocationsMs);
-                }
-            }
-            log.error("Failed completing action [ {} ]. Giving up!", action.getActionName());
-            if (action.getOnFail() != null) {
-                action.getOnFail().run();
-            }
-            return false;
-        }
-    }
 }
