+++ pulsar-functions/utils/src/test/java/org/apache/pulsar/functions/utils/ActionsTest.java
@@ -16,12 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.runtime;
+package org.apache.pulsar.functions.utils;
 
+import org.apache.pulsar.functions.utils.Actions;
 import org.testng.annotations.Test;
 
 import java.util.function.Supplier;
 
+import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -29,23 +31,23 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.testng.Assert.assertEquals;
 
-public class RuntimeUtilsTest {
+public class ActionsTest {
 
     @Test
     public void testActions() throws InterruptedException {
 
         // Test for success
-        Supplier<RuntimeUtils.Actions.ActionResult> supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        Supplier<Actions.ActionResult> supplier1 = mock(Supplier.class);
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
-        Supplier<RuntimeUtils.Actions.ActionResult> supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        Supplier<Actions.ActionResult> supplier2 = mock(Supplier.class);
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
-        Runnable onFail = mock(Runnable.class);
-        Runnable onSucess = mock(Runnable.class);
+        java.util.function.Consumer<Actions.ActionResult> onFail = mock(java.util.function.Consumer.class);
+        java.util.function.Consumer<Actions.ActionResult> onSucess = mock(java.util.function.Consumer.class);
 
-        RuntimeUtils.Actions.Action action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+        Actions.Action action1 = spy(
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(100)
@@ -55,38 +57,38 @@ public class RuntimeUtilsTest {
                         .onSuccess(onSucess)
                         .build());
 
-        RuntimeUtils.Actions.Action action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+        Actions.Action action2 = spy(
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
                         .supplier(supplier2)
                         .build());
 
-        RuntimeUtils.Actions actions = RuntimeUtils.Actions.newBuilder()
+        Actions actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
 
         assertEquals(actions.numActions(), 2);
         verify(supplier1, times(1)).get();
-        verify(onFail, times(0)).run();
-        verify(onSucess, times(1)).run();
+        verify(onFail, times(0)).accept(any());
+        verify(onSucess, times(1)).accept(any());
         verify(supplier2, times(1)).get();
 
         // test only run 1 action
 
         supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
         supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
-        onFail = mock(Runnable.class);
-        onSucess = mock(Runnable.class);
+        onFail = mock(java.util.function.Consumer.class);
+        onSucess = mock(java.util.function.Consumer.class);
 
         action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(100)
@@ -97,7 +99,7 @@ public class RuntimeUtilsTest {
                         .build());
 
         action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
@@ -106,30 +108,30 @@ public class RuntimeUtilsTest {
                         .onSuccess(onSucess)
                         .build());
 
-        actions = RuntimeUtils.Actions.newBuilder()
+        actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
 
         assertEquals(actions.numActions(), 2);
         verify(supplier1, times(1)).get();
-        verify(onFail, times(0)).run();
-        verify(onSucess, times(1)).run();
+        verify(onFail, times(0)).accept(any());
+        verify(onSucess, times(1)).accept(any());
         verify(supplier2, times(0)).get();
 
         // test retry
 
         supplier1 = mock(Supplier.class);
-        when(supplier1.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(false).build());
+        when(supplier1.get()).thenReturn(Actions.ActionResult.builder().success(false).build());
 
         supplier2 = mock(Supplier.class);
-        when(supplier2.get()).thenReturn(RuntimeUtils.Actions.ActionResult.builder().success(true).build());
+        when(supplier2.get()).thenReturn(Actions.ActionResult.builder().success(true).build());
 
-        onFail = mock(Runnable.class);
-        onSucess = mock(Runnable.class);
+        onFail = mock(java.util.function.Consumer.class);
+        onSucess = mock(java.util.function.Consumer.class);
 
         action1 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action1")
                         .numRetries(10)
                         .sleepBetweenInvocationsMs(10)
@@ -140,22 +142,22 @@ public class RuntimeUtilsTest {
                         .build());
 
         action2 = spy(
-                RuntimeUtils.Actions.Action.builder()
+                Actions.Action.builder()
                         .actionName("action2")
                         .numRetries(20)
                         .sleepBetweenInvocationsMs(200)
                         .supplier(supplier2)
                         .build());
 
-        actions = RuntimeUtils.Actions.newBuilder()
+        actions = Actions.newBuilder()
                 .addAction(action1)
                 .addAction(action2);
         actions.run();
 
         assertEquals(actions.numActions(), 2);
-        verify(supplier1, times(10)).get();
-        verify(onFail, times(1)).run();
-        verify(onSucess, times(0)).run();
+        verify(supplier1, times(11)).get();
+        verify(onFail, times(1)).accept(any());
+        verify(onSucess, times(0)).accept(any());
         verify(supplier2, times(1)).get();
 
     }
